mega_downloader: use asyncio.subprocess and code improvements

+++ userbot/modules/mega_downloader.py
@@ -18,17 +18,18 @@
 #  OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 #  ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from subprocess import PIPE, Popen
+from asyncio import create_subprocess_shell as asyncSubprocess
+from asyncio.subprocess import PIPE as asyncPIPE
 
 import re
 import json
 import os
 import time
 import multiprocessing
+import errno
 
 
 from pySmartDL import SmartDL
-from os.path import exists
 from urllib.error import HTTPError
 
 from userbot import CMD_HELP, LOGS
@@ -49,6 +50,18 @@ async def subprocess_run(cmd, megadl):
             f'stderr: {talk[1]}```')
         return
     return talk
+    subproc = await asyncSubprocess(cmd, stdout=asyncPIPE, stderr=asyncPIPE)
+    stdout, stderr = await subproc.communicate()
+    exitCode = subproc.returncode
+    if exitCode != 0:
+        await megadl.edit(
+            '**An error was detected while running subprocess**\n'
+            f'```exit code: {exitCode}\n'
+            f'stdout: {stdout.decode().strip()}\n'
+            f'stderr: {stderr.decode().strip()}```')
+        return exitCode
+    return stdout, stderr
+
 
 
 @register(outgoing=True, pattern=r"^.mega(?: |$)(.*)")
@@ -77,7 +90,7 @@ async def mega_download(url, megadl):
     cmd = f'bin/megadown -q -m {link}'
     result = await subprocess_run(cmd, megadl)
     try:
-        data = json.loads(result[0])
+        data = json.loads(result[0].decode().strip())
     except json.JSONDecodeError:
         await megadl.edit("`Error: Can't extract the link`\n")
         return
@@ -117,6 +130,7 @@ async def mega_download(url, megadl):
                 f" @ {speed}"
                 f"\nETA: {estimated_total_time}"
             )
+
             if status == "Downloading":
                 await megadl.edit(current_message)
                 time.sleep(0.2)
@@ -126,17 +140,29 @@ async def mega_download(url, megadl):
                     display_message = current_message
         except Exception as e:
             LOGS.info(str(e))
+            if status == "Downloading" and display_message != current_message:
+                await megadl.edit(current_message)
+                display_message = current_message
+            elif status == "Combining" and current_message != display_message:
+                await megadl.edit(current_message)
+                display_message = current_message
+        except Exception:
+            pass
+
     if downloader.isSuccessful():
         download_time = downloader.get_dl_time(human=True)
-        if exists(temp_file_name):
+        try:
             P = multiprocessing.Process(target=await decrypt_file(
                 file_name, temp_file_name, hex_key, hex_raw_key, megadl), name="Decrypt_File")
             P.start()
             P.join()
-            if exists(file_name):
-                await megadl.edit(f"`{file_name}`\n\n"
-                                  "Successfully downloaded\n"
-                                  f"Download took: {download_time}")
+        except FileNotFoundError as e:
+            await megadl.edit(str(e))
+            return
+        else:
+            await megadl.edit(f"`{file_name}`\n\n"
+                              "Successfully downloaded\n"
+                              f"Download took: {download_time}")
     else:
         await megadl.edit("Failed to download...")
         for e in downloader.get_errors():
@@ -151,8 +177,11 @@ async def decrypt_file(file_name, temp_file_name,
                        hex_key, hex_raw_key, megadl):
     cmd = ("cat '{}' | openssl enc -d -aes-128-ctr -K {} -iv {} > '{}'"
            .format(temp_file_name, hex_key, hex_raw_key, file_name))
-    await subprocess_run(cmd, megadl)
-    os.remove(temp_file_name)
+    if await subprocess_run(cmd, megadl):
+        os.remove(temp_file_name)
+    else:
+        raise FileNotFoundError(
+            errno.ENOENT, os.strerror(errno.ENOENT), file_name)
     return
